Remove 'pip search'
