+++ docs/html/cli/index.md
@@ -33,13 +33,6 @@ pip_hash
 
 ```{toctree}
 :maxdepth: 1
-:caption: Package Index information
-
-pip_search
-```
-
-```{toctree}
-:maxdepth: 1
 :caption: Managing pip itself
 
 pip_cache
