--- docs/man/commands/search.rst
@@ -1,20 +1,0 @@
-:orphan:
-
-==========
-pip-search
-==========
-
-Description
-***********
-
-.. pip-command-description:: search
-
-Usage
-*****
-
-.. pip-command-usage:: search
-
-Options
-*******
-
-.. pip-command-options:: search
