+++ news/5216.removal.rst
@@ -1,0 +1,1 @@
+Remove the ``pip search`` command as it is basically a no-op with PyPI.
