+++ setup.cfg
@@ -66,7 +66,6 @@ markers =
     svn: VCS: Subversion
     mercurial: VCS: Mercurial
     git: VCS: git
-    search: tests for 'pip search'
 
 [coverage:run]
 branch = True
