+++ tests/conftest.py
@@ -64,21 +64,12 @@ def pytest_addoption(parser: Parser) -> None:
         default=False,
         help="use venv for virtual environment creation",
     )
-    parser.addoption(
-        "--run-search",
-        action="store_true",
-        default=False,
-        help="run 'pip search' tests",
-    )
 
 
 def pytest_collection_modifyitems(config: Config, items: List[pytest.Item]) -> None:
     for item in items:
         if not hasattr(item, "module"):  # e.g.: DoctestTextfile
             continue
-
-        if item.get_closest_marker("search") and not config.getoption("--run-search"):
-            item.add_marker(pytest.mark.skip("pip search test skipped"))
 
         if "CI" in os.environ:
             # Mark network tests as flaky
